Correct name of variables in Car, moved BinarySearch and Searcheble to search and moved  Generic Sorted and FieldSorted to sort
