+++ src/search/Searchable.java
@@ -1,4 +1,4 @@
-package sort;
+package search;
 
 public interface Searchable<T> {
     int binarySearch(T[] array, T key);
