+++ src/sort/GenericSorted.java
@@ -1,4 +1,4 @@
-package search;
+package sort;
 
 public class GenericSorted {
 }
